(feat): Agrega control de permisos en las rutas de categorías y servicios
Se implementan middleware de permisos en las rutas de gestión de categorías y servicios, asegurando que solo los usuarios con los permisos adecuados puedan acceder a las funcionalidades de visualización, creación, edición y eliminación. Esto mejora la seguridad y el control de acceso en la aplicación.

+++ routes/tenants/categories.php
@@ -1,12 +1,13 @@
 <?php
 
+use App\Enums\Permissions;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\CategoryController;
 
 
 Route::group(['prefix' => 'categories'], function () {
-    Route::get('/', [CategoryController::class, 'index'])->name('categories.index');
-    Route::post('/', [CategoryController::class, 'store'])->name('categories.store');
-    Route::put('/{categories}', [CategoryController::class, 'update'])->name('categories.update');
-    Route::delete('/{categories}', [CategoryController::class, 'destroy'])->name('categories.destroy');
+    Route::get('/', [CategoryController::class, 'index'])->name('categories.index')->middleware('can:' . Permissions::CATEGORIES_VIEW);
+    Route::post('/', [CategoryController::class, 'store'])->name('categories.store')->middleware('can:' . Permissions::CATEGORIES_CREATE);
+    Route::put('/{categories}', [CategoryController::class, 'update'])->name('categories.update')->middleware('can:' . Permissions::CATEGORIES_EDIT);
+    Route::delete('/{categories}', [CategoryController::class, 'destroy'])->name('categories.destroy')->middleware('can:' . Permissions::CATEGORIES_DELETE);
 });
